requirement.txt
python-dotenv
python-docx
PyPDF2
openai